pyproject.toml
[project]
name = "eval-acp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "postchain-client-py>=0.2.3",
    "python-dotenv>=1.2.1",
    "virtuals-acp>=0.3.12",
]
